Calibration: convert Float tap fractions to Double for quickCalibrate
Compose Offset.x/y and the ContentScale.Crop transform we mirror are
Float, but FisheyeProjection.quickCalibrate takes Double. Add explicit
.toDouble() at the call site so the Kotlin compiler stops complaining.

+++ app/src/main/java/de/astronarren/allsky/ui/components/CalibrationScreen.kt
@@ -264,8 +264,8 @@ fun CalibrationScreen(
                             }
                             val cal = FisheyeProjection.quickCalibrate(
                                 observedBody = tgt.coords,
-                                tappedPxFrac = imgX / img.width,
-                                tappedPyFrac = imgY / img.height,
+                                tappedPxFrac = (imgX / img.width).toDouble(),
+                                tappedPyFrac = (imgY / img.height).toDouble(),
                             )
                             scope.launch {
                                 userPreferences.saveFisheyeCalibration(cal)
